requirements.txt
alembic==1.16.4
attrs==25.3.0
babel==2.17.0
bidict==0.23.1
blinker==1.9.0
cached-property==2.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
configparser==7.2.0
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
parsedatetime==2.6
pathlib==1.0.1
poker==0.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
requests==2.32.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.42
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.interface==7.2
